refactor: rename state.particles -> properties
