+++ expertsystem/state/properties.py
